+++ src/main/java/sky/Ui.java
@@ -1,3 +1,5 @@
+package sky;
+
 import java.util.Scanner;
 
 /**
