+++ src/main/java/sky/command/Command.java
@@ -1,3 +1,10 @@
+package sky.command;
+
+import sky.Storage;
+import sky.TaskList;
+import sky.exception.TextNoMeaningException;
+import sky.Ui;
+
 /**
  * The Command class deals with executing tasks involving TaskList, Ui, and Storage.
  */
