+++ src/main/java/sky/command/InvalidCommand.java
@@ -1,3 +1,10 @@
+package sky.command;
+
+import sky.Storage;
+import sky.TaskList;
+import sky.exception.TextNoMeaningException;
+import sky.Ui;
+
 /**
  * The InvalidCommand class deals with invalid inputs.
  */
