+++ src/main/java/sky/task/Deadline.java
@@ -1,3 +1,5 @@
+package sky.task;
+
 public class Deadline extends Task {
     protected String by;
