+++ src/main/java/sky/task/Task.java
@@ -1,3 +1,5 @@
+package sky.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
